// src/api/generated/model/apiResponseExamSessionStatistics.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { ExamSessionStatistics } from "./examSessionStatistics";

export interface ApiResponseExamSessionStatistics {
  success?: boolean;
  message?: string;
  data?: ExamSessionStatistics;
  errors?: string[];
  timestamp?: string;
  path?: string;
}