// src/api/generated/model/updateExamSessionStateRequest.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateExamSessionStateRequestSessionState } from "./updateExamSessionStateRequestSessionState";

export interface UpdateExamSessionStateRequest {
  /** @minLength 1 */
  examSessionId: string;
  sessionState: UpdateExamSessionStateRequestSessionState;
}